Revert "Ability to read Excel files"
This reverts commit ef7c9343d6a307cd43c62dbd10998d00fc5f4936.
