+++ tests/spark/transformations/test_sql_transform.py
@@ -1,11 +1,10 @@
 from textwrap import dedent
 
 import pytest
+from tests.conftest import TEST_DATA_PATH
 
 from koheesio.logger import LoggingFactory
 from koheesio.spark.transformations.sql_transform import SqlTransform
-
-from tests.conftest import TEST_DATA_PATH
 
 pytestmark = pytest.mark.spark
