+++ tests/utils/test_utils.py
@@ -25,6 +25,26 @@ def test_import_class():
     assert import_class("datetime.datetime") == datetime.datetime
 
 
+@pytest.mark.parametrize(
+    "env_var_value, expected_result",
+    [("lts_11_spark_3_scala_2.12", True), ("unit_test", True), (None, False)],
+)
+def test_on_databricks(env_var_value, expected_result):
+    if env_var_value is not None:
+        with patch.dict(os.environ, {"DATABRICKS_RUNTIME_VERSION": env_var_value}):
+            assert on_databricks() == expected_result
+    else:
+        with patch.dict(os.environ, clear=True):
+            assert on_databricks() == expected_result
+
+
+def test_schema_struct_to_schema_str():
+    struct_schema = StructType([StructField("a", StringType()), StructField("b", StringType())])
+    val = schema_struct_to_schema_str(struct_schema)
+    assert val == "a STRING,\nb STRING"
+    assert schema_struct_to_schema_str(None) == ""
+
+
 def test_get_random_string():
     assert get_random_string(10) != get_random_string(10)
     assert len(get_random_string(10)) == 10
